requirement.txt
requests
selenium
datetime
PyQt6
pyinstaller
python-dateutil